kernel: move syscall constructor to arch helper
How a system call is created from registers is ABI-specific. We should
use the trd-specific function based on the architecture.
