+++ arch/riscv/src/syscall.rs
@@ -654,13 +654,14 @@ impl kernel::syscall::UserspaceKernelBoundary for SysCall {
                         // instruction. The hardware does not do this for us.
                         state.pc = state.pc.wrapping_add(4);
 
-                        let syscall = kernel::syscall::Syscall::from_register_arguments(
-                            state.regs[R_A4] as u8,
-                            state.regs[R_A0] as usize,
-                            (state.regs[R_A1] as usize).into(),
-                            (state.regs[R_A2] as usize).into(),
-                            (state.regs[R_A3] as usize).into(),
-                        );
+                        let syscall =
+                            kernel::utilities::arch_helpers::syscall_from_register_arguments_trd104(
+                                state.regs[R_A4] as u8,
+                                state.regs[R_A0] as usize,
+                                (state.regs[R_A1] as usize).into(),
+                                (state.regs[R_A2] as usize).into(),
+                                (state.regs[R_A3] as usize).into(),
+                            );
 
                         match syscall {
                             Some(s) => ContextSwitchReason::SyscallFired { syscall: s },
